requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
